requirements.txt
python-telegram-bot
google_spreadsheet
google-auth-oauthlib